fix zmq.serving error on windows
